Add Scribus source file (June 2020, November 2021)
The two pages of this article were initially created in the context
of a larger document, in June 2020. In November 2021, I extracted
the two pages to make the article standalone. I added the date and
author name. I also reworked the logo Beauty & the Bricks to match
my original concept, with the end of the tail of the y overlapping
the dot of the i in the word Bricks below.